adding casts and reactions streams and code refactor
