+++ apps/replicator/src/processors/index.ts
@@ -43,6 +43,7 @@ import {
   farcasterTimeToDate,
   StoreMessageOperation,
   toHexEncodedUint8Array,
+  putKinesisRecords
 } from "../util.js";
 import { processOnChainEvent } from "./onChainEvent.js";
 import { processUserNameProofAdd, processUserNameProofMessage, processUserNameProofRemove } from "./usernameProof.js";
@@ -58,37 +59,6 @@ import AWS from "aws-sdk";
 import { Records } from "aws-sdk/clients/rdsdataservice.js";
 
 
-// const credentials = new AWS.Credentials({
-//   accessKeyId: AWS_ACCESS_KEY_ID,
-//   secretAccessKey: AWS_SECRET_ACCESS_KEY
-// });
-
-AWS.config.update({
-  region: "eu-west-1" 
-});
-const kinesis = new AWS.Kinesis();
-
-interface KinesisRecord {
-  Data: string;
-  PartitionKey: string;
-}
-
-async function putKinesisRecords(records: KinesisRecord[]) {
-  const params = {
-    Records: records,
-    StreamName: "farcaster-stream", // Replace 'your-stream-name' with your Kinesis stream name
-  };
-
-  // Put records into the Kinesis stream
-  kinesis.putRecords(params, (err, data) => {
-    if (err) {
-      console.error("Error putting records:", err);
-    } else {
-      console.log(data);
-      console.log("Successfully put records:", data.Records.length);
-    }
-  });
-}
 
 export async function processOnChainEvents(events: OnChainEvent[], db: DB, log: Logger, redis: Redis, isHubEvent: boolean = false) {
   for (const event of events) {
@@ -375,47 +345,6 @@ function transformMessage(message: Message): Message {
   return message;
 }
 
-// export async function storeMessage(
-//   message: Message,
-//   operation: StoreMessageOperation,
-//   trx: DBTransaction,
-//   log: Logger,
-// ): Promise<void>  {
-//   if (!message.data) throw new Error("Message missing data!"); // Shouldn't happen
-//   const now = new Date();
-
-//   log.debug(`Storing message ${bytesToHex(message.hash)} via ${operation} operation`);
-//   let records = [];
-    
-//   let recordsJson = {
-//     createdAt: now,
-//     updatedAt: now,
-//     fid: message.data.fid,
-//     type: message.data.type,
-//     timestamp: farcasterTimeToDate(message.data.timestamp),
-//     hash: message.hash,
-//     hashScheme: message.hashScheme,
-//     signature: message.signature,
-//     signatureScheme: message.signatureScheme,
-//     signer: message.signer,
-//     raw: Message.encode(message).finish(),
-//     deletedAt: operation === "delete" ? now : null,
-//     prunedAt: operation === "prune" ? now : null,
-//     revokedAt: operation === "revoke" ? now : null,
-//     body: JSON.stringify(convertProtobufMessageBodyToJson(message)),
-//   }
-  
-//   records = [
-//     {
-//       Data: JSON.stringify(recordsJson),
-//       PartitionKey: "MESSAGE_ADD",
-//     },
-//   ];
-//   console.log(`push kinesis start`);
-//   await putKinesisRecords(records);
-//   console.log(`push kinesis end`);
-// }
-
 export async function storeMessage(
   message: Message,
   operation: StoreMessageOperation,
@@ -454,7 +383,7 @@ export async function storeMessage(
     },
   ];
   // console.log(`push kinesis start`);
-  // await putKinesisRecords(records);
+  // await putKinesisRecords(records, "farcaster-stream");
   // console.log(`push kinesis end`);
   await execute(
     trx
