+++ apps/replicator/src/processors/reaction.ts
@@ -7,46 +7,13 @@ import {
 } from "@farcaster/hub-nodejs";
 import { Selectable, sql } from "kysely";
 import { buildAddRemoveMessageProcessor } from "../messageProcessor.js";
-import { bytesToHex, farcasterTimeToDate, isAfterTargetTimeToday, isBetweenPeriod, isToday } from "../util.js";
+import { bytesToHex, farcasterTimeToDate, isAfterTargetTimeToday, isBetweenPeriod, isToday, putKinesisRecords } from "../util.js";
 import { ReactionRow, executeTakeFirst } from "../db.js";
 import { AssertionError, HubEventProcessingBlockedError } from "../error.js";
 import AWS from "aws-sdk";
 import { Records } from "aws-sdk/clients/rdsdataservice.js";
 import {AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY, useTimePeriodKinesis } from "../env.js";
 
-
-// const credentials = new AWS.Credentials({
-//   accessKeyId: AWS_ACCESS_KEY_ID,
-//   secretAccessKey: AWS_SECRET_ACCESS_KEY
-// });
-
-AWS.config.update({
-  region: "eu-west-1" 
-});
-
-const kinesis = new AWS.Kinesis();
-
-interface KinesisRecord {
-  Data: string;
-  PartitionKey: string;
-}
-
-async function putKinesisRecords(records: KinesisRecord[]) {
-  const params = {
-    Records: records,
-    StreamName: "farcaster-stream", // Replace 'your-stream-name' with your Kinesis stream name
-  };
-
-  // Put records into the Kinesis stream
-  kinesis.putRecords(params, (err, data) => {
-    if (err) {
-      console.error("Error putting records:", err);
-    } else {
-      console.log(data);
-      console.log("Successfully put records:", data.Records.length);
-    }
-  });
-}
 
 
 const { processAdd, processRemove } = buildAddRemoveMessageProcessor<
@@ -175,7 +142,14 @@ const { processAdd, processRemove } = buildAddRemoveMessageProcessor<
     // if (isAfterTargetTimeToday(farcasterTimeToDate(message.data.timestamp)) || (useTimePeriodKinesis && isBetweenPeriod(farcasterTimeToDate(message.data.timestamp)))) {
     if(isToday(farcasterTimeToDate(message.data.timestamp))) {
       console.log(`push kinesis start`);
-      await putKinesisRecords(records);
+      await putKinesisRecords(records, "farcaster-stream");
+      records = [
+        {
+          Data: JSON.stringify(recordsJson),
+          PartitionKey: bytesToHex(targetCastId?.hash),
+        },
+      ];
+      await putKinesisRecords(records, "farcaster-reactions-stream");
       console.log(`push kinesis end`);
     }
