+++ apps/replicator/src/processors/verification.ts
@@ -2,46 +2,13 @@ import { MessageType, Protocol, VerificationAddAddressMessage, VerificationRemov
 import { Selectable, sql } from "kysely";
 import { buildAddRemoveMessageProcessor } from "../messageProcessor.js";
 import { executeTakeFirst, executeTakeFirstOrThrow, VerificationRow } from "../db.js";
-import { bytesToHex, farcasterTimeToDate, StoreMessageOperation } from "../util.js";
+import { bytesToHex, farcasterTimeToDate, StoreMessageOperation, putKinesisRecords } from "../util.js";
 import base58 from "bs58";
 
 import AWS from "aws-sdk";
 import { Records } from "aws-sdk/clients/rdsdataservice.js";
 import {AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY } from "../env.js";
 
-
-// const credentials = new AWS.Credentials({
-//   accessKeyId: AWS_ACCESS_KEY_ID,
-//   secretAccessKey: AWS_SECRET_ACCESS_KEY
-// });
-
-AWS.config.update({
-  region: "eu-west-1" 
-});
-
-const kinesis = new AWS.Kinesis();
-
-interface KinesisRecord {
-  Data: string;
-  PartitionKey: string;
-}
-
-async function putKinesisRecords(records: KinesisRecord[]) {
-  const params = {
-    Records: records,
-    StreamName: "farcaster-stream", // Replace 'your-stream-name' with your Kinesis stream name
-  };
-
-  // Put records into the Kinesis stream
-  kinesis.putRecords(params, (err, data) => {
-    if (err) {
-      console.error("Error putting records:", err);
-    } else {
-      console.log(data);
-      console.log("Successfully put records:", data.Records.length);
-    }
-  });
-}
 
 
 const { processAdd: processAddEthereum, processRemove } = buildAddRemoveMessageProcessor<
@@ -130,7 +97,7 @@ const { processAdd: processAddEthereum, processRemove } = buildAddRemoveMessageP
       },
     ];
     // console.log(`push kinesis start`);
-    // await putKinesisRecords(records);
+    // await putKinesisRecords(records, "farcaster-stream");
     // console.log(`push kinesis end`);
 
     // Upsert the verification, if it's shadowed by a remove, mark it as deleted
@@ -159,44 +126,6 @@ const { processAdd: processAddEthereum, processRemove } = buildAddRemoveMessageP
     // No-op
   },
 });
-
-
-// async function processVerificationsAddKinesis(message: Message,
-//   deleted: StoreMessageOperation,
-// ): Promise<void>  {
-//   const {
-//     data: { fid, verificationAddAddressBody: verificationAddBody },
-//   } = message;
-
-//   const timestamp = farcasterTimeToDate(message.data.timestamp);
-
-//   const updatedProps = {
-//     timestamp,
-//     hash: message.hash,
-//     signerAddress: verificationAddBody.address,
-//     blockHash: verificationAddBody.blockHash,
-//     signature: verificationAddBody.claimSignature,
-//   };
-  
-//   let records = [];
-  
-//   let recordsJson = {
-//     deletedAt: deleted ? new Date() : null,
-//     fid,
-//     ...updatedProps,
-//   }
-  
-//   records = [
-//     {
-//       Data: JSON.stringify(recordsJson),
-//       PartitionKey: "VERIFICATIONS_ADD",
-//     },
-//   ];
-//   console.log(`push kinesis start`);
-//   await putKinesisRecords(records);
-//   console.log(`push kinesis end`);
-// };
-
 
 
 // TODO: implement processAdd support for Solana in separate PR
